+++ package.json
@@ -2,7 +2,7 @@
   "name": "custom-launcher",
   "version": "0.0.4",
   "description": "Launcher de Minecraft personalizable basado en ElectronJS",
-  "main": "src/javascript/main.js",
+  "main": "src/main.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
     "start": "electron ."
