+++ src/js/json/settings.json
